+++ include/openspace/interaction/navigationhandler.h
